+++ Events/CarrierDecommissionCancelledEvent.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using EddiDataDefinitions;
+using System;
 using Utilities;
 
 namespace EddiEvents
@@ -10,12 +11,16 @@ namespace EddiEvents
         public const string DESCRIPTION = "Triggered when you cancel the decommissioning of your fleet carrier";
         public const string SAMPLE = "{ \"timestamp\":\"2020-03-11T15:12:38Z\", \"event\":\"CarrierCancelDecommission\", \"CarrierID\":3700005632 }";
 
-        // Not intended to be user facing
-        public long? carrierID { get; private set; }
+        [PublicAPI( "The carrier's numeric ID" )]
+        public long carrierID { get; private set; }
 
-        public CarrierDecommissionCancelledEvent(DateTime timestamp, long? carrierId) : base(timestamp, NAME)
+        [PublicAPI( "The carrier type (e.g. Fleet Carrier or Squadron Carrier), as an object with 'localizedName' and 'invariantName' properties" )]
+        public StationModel carrierType { get; private set; }
+
+        public CarrierDecommissionCancelledEvent ( DateTime timestamp, long carrierId, StationModel carrierType ) : base(timestamp, NAME)
         {
-            carrierID = carrierId;
+            this.carrierID = carrierId;
+            this.carrierType = carrierType;
         }
     }
 }