+++ Events/CarrierPadsLockedEvent.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using EddiDataDefinitions;
+using System;
 using Utilities;
 
 namespace EddiEvents
@@ -8,15 +9,18 @@ namespace EddiEvents
     {
         public const string NAME = "Carrier pads locked";
         public const string DESCRIPTION = "Triggered when your fleet carrier locks landing pads prior to jumping";
-        public static CarrierPadsLockedEvent SAMPLE = new CarrierPadsLockedEvent(DateTime.UtcNow, 3700357376);
+        public static CarrierPadsLockedEvent SAMPLE = new CarrierPadsLockedEvent(DateTime.UtcNow, 3700357376, StationModel.FleetCarrier );
 
-        // These properties are not intended to be user facing
+        [PublicAPI( "The carrier's numeric ID" )]
+        public long carrierID { get; private set; }
 
-        public long carrierId { get; private set; }
+        [PublicAPI( "The carrier type (e.g. Fleet Carrier or Squadron Carrier), as an object with 'localizedName' and 'invariantName' properties" )]
+        public StationModel carrierType { get; private set; }
 
-        public CarrierPadsLockedEvent(DateTime timestamp, long carrierId) : base(timestamp, NAME)
+        public CarrierPadsLockedEvent ( DateTime timestamp, long carrierId, StationModel carrierType ) : base(timestamp, NAME)
         {
-            this.carrierId = carrierId;
+            this.carrierID = carrierId;
+            this.carrierType = carrierType;
         }
     }
 }